Use most recent enum options
In some cases there may be more than one list of options for enumerated values
based on the version of ParFlow. Down the road this selection should be based on
the version set by the user, but until that feature is added we will default to
the most recent version listed.

Signed-off-by: Brianna Major <brianna.major@kitware.com>